fix: Assign current user in prepare_group_membership
A bool was being assigned to compare_user instead of the appropiate
User object. As there could also be an issue if the member_of
attribute is None, which would have required another if check, I went
with a try block to handle both extra cases.

There were also typos in a few attribute names.